refactored Attribute import
